Fix review findings

+++ src/cloud_server/cloud_server.h
@@ -23,8 +23,8 @@ struct cloud_data_struct
 	struct gps_data_struct gps_data;
 	struct client_controller_data_struct client_controller_data;
 	struct can_data_struct can_data;
-	float service_time;
-	int distance_travelled;
+	int service_time;
+	float distance_travelled;
 };
 
 void *write_to_cloud(void *); /* pthread to display all serial data */
